- fixed a couple of errors stopping the program compiling in the windows 10 ubuntu bash shell - replaced long/short types with int32_t and int16_t to fix issues when compiling 64-bit versions

+++ Image.cpp
@@ -21,20 +21,20 @@ bool Image::SetPixelColour(int x, int y, float r, float g, float b)
 
 struct BitmapHeader
 {
-	long size_of_file;
-	long reserve;
-	long offset_of_pixel_data;
-	long size_of_header;
-	long width;
-	long height;
-	short num_of_colour_plane;
-	short num_of_bit_per_pix;
-	long compression;
-	long size_of_pix_data;
-	long h_resolution;
-	long v_resolution;
-	long num_of_colour_in_palette;
-	long important_colours;
+	int32_t size_of_file;
+	int32_t reserve;
+	int32_t offset_of_pixel_data;
+	int32_t size_of_header;
+	int32_t width;
+	int32_t height;
+	int16_t num_of_colour_plane;
+	int16_t num_of_bit_per_pix;
+	int32_t compression;
+	int32_t size_of_pix_data;
+	int32_t h_resolution;
+	int32_t v_resolution;
+	int32_t num_of_colour_in_palette;
+	int32_t important_colours;
 } ;
 
 bool Image::SaveImage(const char* outputName)
@@ -48,7 +48,7 @@ bool Image::SaveImage(const char* outputName)
 		int rowLenInBytes =  ((m_w * m_bpp) + xPad);
 		BitmapHeader header;
 		header.size_of_file =  sizeof(header) + sizeof(BM) + (rowLenInBytes * m_h);
-		header.reserve = 0000; 
+		header.reserve = 0000;
 		header.offset_of_pixel_data = 54;
 		header.size_of_header = 40;
 		header.width = m_w;
@@ -64,7 +64,7 @@ bool Image::SaveImage(const char* outputName)
 
 		ofstream file;
 		file.open (outputName, ios::out | ios::trunc | ios::binary);
-		
+
 		if (file.is_open())
 		{
 			file.write ((char*)(&BM), 2);
@@ -74,7 +74,7 @@ bool Image::SaveImage(const char* outputName)
 			{
 				int rowLen = (m_w * m_bpp);
 				for (int x=0;x<m_w;x++)
-				{	
+				{
 					int buffPos = (x*m_bpp)+(rowLen * y);
 					file.write((char*)&m_imageData[buffPos+2], 1);
 					file.write((char*)&m_imageData[buffPos+1], 1);
